issue 4 - rename connectioncontext to connectionfactory and close on job runs, r=twobraids
